# Improve performance of RuntimeHelpers.GetMethodTable