--- a/src/coreclr/src/System.Private.CoreLib/src/System/Runtime/CompilerServices/RuntimeHelpers.CoreCLR.cs
+++ b/src/coreclr/src/System.Private.CoreLib/src/System/Runtime/CompilerServices/RuntimeHelpers.CoreCLR.cs
@@ -260,22 +260,13 @@ namespace System.Runtime.CompilerServices
         // GC.KeepAlive(o);
         //
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [Intrinsic]
         internal static unsafe MethodTable* GetMethodTable(object obj)
         {
-            Debug.Assert(obj != null);
-
-            // We know that the first data field in any managed object is immediately after the
-            // method table pointer, so just back up one pointer and immediately deref.
-            // This is not ideal in terms of minimizing instruction count but is the best we can do at the moment.
+            // The body of this function will be replaced by the EE with unsafe code
+            // See getILIntrinsicImplementationForRuntimeHelpers for how this happens.
 
             return (MethodTable *)Unsafe.Add(ref Unsafe.As<byte, IntPtr>(ref obj.GetRawData()), -1);
-
-            // The JIT currently implements this as:
-            // lea tmp, [rax + 8h] ; assume rax contains the object reference, tmp is type IntPtr&
-            // mov tmp, qword ptr [tmp - 8h] ; tmp now contains the MethodTable* pointer
-            //
-            // Ideally this would just be a single dereference:
-            // mov tmp, qword ptr [rax] ; rax = obj ref, tmp = MethodTable* pointer
         }
     }