--- a/src/coreclr/src/vm/jitinterface.cpp
+++ b/src/coreclr/src/vm/jitinterface.cpp
@@ -7485,6 +7485,48 @@ bool getILIntrinsicImplementationForRuntimeHelpers(MethodDesc * ftn,
         return true;
     }
 
+    if (tk == MscorlibBinder::GetMethod(METHOD__RUNTIME_HELPERS__GET_METHOD_TABLE)->GetMemberDef())
+    {
+        mdToken tokRawData = MscorlibBinder::GetField(FIELD__RAW_DATA__DATA)->GetMemberDef();
+
+        // In the CLR, an object is laid out as follows.
+        // [ object_header || MethodTable* (64-bit pointer) || instance_data ]
+        //                    ^                                ^-- ref <theObj>.firstField points here
+        //                    `-- <theObj> reference (type O) points here
+        //
+        // So essentially what we want to do is to turn an object reference (type O) into a
+        // native int&, then dereference it to get the MethodTable*. (Essentially, an object
+        // reference is a MethodTable**.) Per ECMA-335, Sec. III.1.5, we can add
+        // (but not subtract) a & and an int32 to produce a &. So we'll get a reference to
+        // <theObj>.firstField (type &), then back up one pointer length to get a value of
+        // essentially type (MethodTable*)&. Both of these are legal GC-trackable references
+        // to <theObj>, regardless of <theObj>'s actual length.
+
+        static BYTE ilcode[] = { CEE_LDARG_0,         // stack contains [ O ] = <theObj>
+                                 CEE_LDFLDA,0,0,0,0,  // stack contains [ & ] = ref <theObj>.firstField
+#ifdef _TARGET_64BIT_
+                                 CEE_LDC_I4_8,        // stack contains [ &, int32 ]
+#else
+                                 CEE_LDC_I4_4,        // stack contains [ &, int32 ]
+#endif
+                                 CEE_NEG,             // stack contains [ &, int32 ]
+                                 CEE_ADD,             // stack contains [ & ] = ref <theObj>.methodTablePtr
+                                 CEE_LDIND_I,         // stack contains [ native int ] = <theObj>.methodTablePtr
+                                 CEE_RET };
+
+        ilcode[2] = (BYTE)(tokRawData);
+        ilcode[3] = (BYTE)(tokRawData >> 8);
+        ilcode[4] = (BYTE)(tokRawData >> 16);
+        ilcode[5] = (BYTE)(tokRawData >> 24);
+
+        methInfo->ILCode = const_cast<BYTE*>(ilcode);
+        methInfo->ILCodeSize = sizeof(ilcode);
+        methInfo->maxStack = 2;
+        methInfo->EHcount = 0;
+        methInfo->options = (CorInfoOptions)0;
+        return true;
+    }
+
     if (tk == MscorlibBinder::GetMethod(METHOD__RUNTIME_HELPERS__ENUM_EQUALS)->GetMemberDef())
     {
         // Normally we would follow the above pattern and unconditionally replace the IL,