… src/coreclr/src/vm/mscorlib.h