Some small perf improvements to `RuntimeHelpers.GetMethodTable`. Previously the codegen looked like:

```asm
488d5108        lea     rdx,[rcx+8]
488b52f8        mov     rdx,qword ptr [rdx-8]
```

Now the codegen looks like:

```asm
488b11          mov     rdx,qword ptr [rcx]   ; no intermediate 'lea' instruction, saving 5 bytes of codegen
```

Below is the microbenchmark for pulling a `Span<byte>` from a `Memory<byte>` backed by a `byte[]`.

|                      Method | Toolchain |     Mean |     Error |    StdDev | Ratio |
|---------------------------- |---------- |---------:|----------:|----------:|------:|
| GetSpan_MemOfBytesFromArray |    master | 1.980 ms | 0.0169 ms | 0.0150 ms |  1.00 |
| GetSpan_MemOfBytesFromArray |  memproto | 1.898 ms | 0.0303 ms | 0.0284 ms |  0.96 |

It's not much, but it could help make a small dent if these APIs are a hot path in a loop.