@AndyAyersMS 2019-11-25
I don't think it is all that common. Most of the code the JIT sees won't use `Unsafe` and explicit address manipulation. Regular field accesses and such are handled reasonably well, even for things like arrays of structs.

However, code that ends up using `Unsafe` and address manipulation is often trying to squeeze out the last bit of perf, so it's a bit of a disconnect to see that the JIT doesn't cope with this as well as it could.

Address mode formation is not as robust in the JIT as it is in native compilers and there are numerous open issues (eg dotnet/coreclr#5526, dotnet/coreclr#10737, dotnet/coreclr#11458, ...) with suggestions on how to improve. But none of these are small work items.