@GrabYourPitchforks 2019-11-26
In the latest iteration I implemented an IL intrinsic rather than declaring new types. Let me know if this is closer to what you're looking for.

I confirmed the codegen is the same as the original iteration: no intermediate `lea` instruction, just a simple `mov rdx, qword ptr [rcx]`.