@GrabYourPitchforks 2019-11-26
Something went horribly wrong with my repo and I can no longer make code changes to this PR. 😬
Abandoning this and opened https://github.com/dotnet/runtime/pull/275 instead. Sorry for the inconvenience.